refactor(stats): extracting info from orders in progress

+++ src/main/java/com/proyect/masterdata/dto/StatsCardDTO.java
@@ -11,12 +11,11 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class StatsCardDTO
-{
+public class StatsCardDTO {
     BigDecimal totalSales;
     Integer totalOrders;
     Integer totalProducts;
-    BigDecimal averageSaleProduct;
+    double averageSaleProduct;
     BigDecimal averageTicket;
     BigDecimal totalDeliveryAmountOrders;
     String orderStatus;
